+++ Runtime/Scripts/DracoNative.cs
@@ -13,10 +13,6 @@
 // limitations under the License.
 //
 
-#if UNITY_2020_2_OR_NEWER
-#define DRACO_MESH_DATA
-#endif
-
 using System;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
@@ -111,18 +107,9 @@ namespace Draco {
 
         bool isPointCloud;
 
-#if DRACO_MESH_DATA
         Mesh.MeshData mesh;
         int indicesCount;
         
-#else
-        Mesh mesh;
-        int streamCount;
-        NativeIndexBufferBase indices;
-        NativeArray<byte>[] vData;
-        byte*[] vDataPtr;
-#endif
-
 #region BlendHack
         // TODO: BLENDHACK; Unity does not support setting bone weights and indices via new Mesh API
         // https://fogbugz.unity3d.com/default.asp?1320869_7g7qeq40va98n6h6
@@ -136,16 +123,12 @@ namespace Draco {
 #endregion BlendHack
 
         public DracoNative(
-#if DRACO_MESH_DATA
             Mesh.MeshData mesh,
-#endif
             bool convertSpace = true
             )
         {
             this.convertSpace = convertSpace;
-#if DRACO_MESH_DATA
             this.mesh = mesh;
-#endif
         }
 
         public JobHandle Init(IntPtr encodedData, int size) {
@@ -212,9 +195,15 @@ namespace Draco {
 
                     DracoAttribute* attribute = null;
                     if (GetAttributeByType(draco, attributeType, i, &attribute)) {
-                        var format = GetVertexAttributeFormat((DataType)attribute->dataType, normalized);
+                        var format = GetVertexAttributeFormat(
+                            (DataType)attribute->dataType, normalized);
                         if (!format.HasValue) { continue; }
-                        var map = new AttributeMap(attribute, type.Value, format.Value, convertSpace && ConvertSpace(type.Value));
+                        var map = new AttributeMap(
+                            attribute,
+                            type.Value,
+                            format.Value,
+                            convertSpace && ConvertSpace(type.Value)
+                            );
                         attributes.Add(map);
                         attributeTypes.Add(type.Value);
                         foundAttribute = true;
@@ -336,9 +325,6 @@ namespace Draco {
                         streamIndex = hasTexCoordOrColor ? 2 : 1;
                         break;
                 }
-#if !DRACO_MESH_DATA
-                streamCount = Mathf.Max(streamCount, streamIndex+1);
-#endif
                 var elementSize = attributeMap.elementSize;
                 attributeMap.offset = streamStrides[streamIndex];
                 attributeMap.stream = streamIndex;
@@ -348,28 +334,6 @@ namespace Draco {
             attributes.Sort();
             Profiler.EndSample(); // CalculateVertexParams
         }
-
-#if !DRACO_MESH_DATA
-        void AllocateIndices(DracoMesh* dracoMesh) {
-            Profiler.BeginSample("AllocateIndices");
-            if (dracoMesh->indexFormat == IndexFormat.UInt16) {
-                indices = new NativeIndexBuffer<ushort>(dracoMesh->numFaces * 3, allocator);
-            } else {
-                indices = new NativeIndexBuffer<uint>(dracoMesh->numFaces * 3, allocator);
-            }
-            Profiler.EndSample(); // AllocateIndices
-        }
-        
-        void AllocateVertexBuffers(DracoMesh* dracoMesh) {
-            int streamIndex;
-            vData = new NativeArray<byte>[streamCount];
-            vDataPtr = new byte*[streamCount];
-            for (streamIndex = 0; streamIndex < streamCount; streamIndex++) {
-                vData[streamIndex] = new NativeArray<byte>(streamStrides[streamIndex] * dracoMesh->numVertices, allocator, NativeArrayOptions.UninitializedMemory);
-                vDataPtr[streamIndex] = (byte*)vData[streamIndex].GetUnsafePtr();
-            }
-        }
-#endif
 
 #if UNITY_EDITOR
         public void DecodeVertexDataSync() {
@@ -399,12 +363,7 @@ namespace Draco {
                 dracoTempResources = dracoTempResources,
                 flip = convertSpace,
                 dataType = mesh.indexFormat == IndexFormat.UInt16 ? DataType.DT_UINT16 : DataType.DT_UINT32, 
-#if DRACO_MESH_DATA
                 mesh = mesh
-#else
-                indicesPtr = indices.unsafePtr,
-                indicesLength =  indices.Length
-#endif
             };
             var jobCount = attributes.Count + 1;
             
@@ -436,13 +395,10 @@ namespace Draco {
                         attribute = map.dracoAttribute,
                         stride = streamStrides[map.stream],
                         flip = map.convertSpace,
-#if DRACO_MESH_DATA                        
+                        componentStride = map.numComponents,
                         mesh = mesh, 
                         streamIndex = map.stream, 
                         offset = map.offset
-#else
-                        dstPtr = vDataPtr[map.stream] + map.offset
-#endif                        
                     };
                     jobHandles[jobIndex] = job.Schedule(decodeVerticesJobHandle);
                 }
@@ -452,12 +408,9 @@ namespace Draco {
                         dracoTempResources = dracoTempResources,
                         attribute = map.dracoAttribute,
                         flip = map.convertSpace,
-#if DRACO_MESH_DATA                        
+                        componentStride = map.numComponents,
                         mesh = mesh, 
                         streamIndex = map.stream
-#else
-                        dstPtr = vDataPtr[map.stream] + map.offset
-#endif
                     };
                     jobHandles[jobIndex] = job.Schedule(decodeVerticesJobHandle);
                 }
@@ -525,11 +478,7 @@ namespace Draco {
             
             Profiler.BeginSample("SetParameters");
             isPointCloud = dracoMesh->isPointCloud;
-#if DRACO_MESH_DATA
             indicesCount = dracoMesh->numFaces * 3;
-#else
-            mesh = new Mesh();
-#endif
             if (!isPointCloud) {
                 mesh.SetIndexBufferParams(dracoMesh->numFaces*3, dracoMesh->indexFormat);
             }
@@ -538,10 +487,6 @@ namespace Draco {
                 vertexParams.Add(map.GetVertexAttributeDescriptor());
             }
             mesh.SetVertexBufferParams(dracoMesh->numVertices, vertexParams.ToArray());
-#if !DRACO_MESH_DATA
-            AllocateIndices(dracoMesh);
-            AllocateVertexBuffers(dracoMesh);
-#endif
             if (hasBoneWeightData) {
                 var boneCount = boneIndexMap.numComponents;
                 bonesPerVertex = new NativeArray<byte>(dracoMesh->numVertices, Allocator.Persistent);
@@ -556,11 +501,7 @@ namespace Draco {
             dracoTempResources.Dispose();
         }
 
-#if DRACO_MESH_DATA
         public bool 
-#else
-        public Mesh
-#endif
         PopulateMeshData() {
             
             Profiler.BeginSample("PopulateMeshData");
@@ -573,52 +514,17 @@ namespace Draco {
             Profiler.BeginSample("MeshAssign");
 
             const MeshUpdateFlags flags = DracoMeshLoader.defaultMeshUpdateFlags;
-
-#if !DRACO_MESH_DATA
-            for (var streamIndex = 0; streamIndex < streamCount; streamIndex++) {
-                mesh.SetVertexBufferData(vData[streamIndex], 0, 0, vData[streamIndex].Length, streamIndex, flags);
-            }
-
-            indices.SetMeshIndexBufferData(mesh);
-            var indicesCount = indices.Length;
-
-            if (hasBoneWeightData) {
-                mesh.SetBoneWeights(bonesPerVertex,boneWeights);
-            }
-#endif
-
+            
             mesh.subMeshCount = 1;
             var submeshDescriptor = new SubMeshDescriptor(0, indicesCount, isPointCloud ? MeshTopology.Points : MeshTopology.Triangles) { firstVertex = 0, baseVertex = 0, vertexCount = mesh.vertexCount };
             mesh.SetSubMesh(0, submeshDescriptor, flags);
             Profiler.EndSample(); // CreateUnityMesh.CreateMesh
-
-#if !DRACO_MESH_DATA
-            Profiler.BeginSample("Dispose");
-            indices.Dispose();
-            foreach (var nativeArray in vData) {
-                nativeArray.Dispose();
-            }
-            if (hasBoneWeightData) {
-                DisposeBoneWeightData();
-            }
-            Profiler.EndSample();
-#endif
             Profiler.EndSample();
             
-#if DRACO_MESH_DATA
             return true;
-#else
-            return mesh;
-#endif
         }
 
         public void DisposeBoneWeightData() {
-#if !DRACO_MESH_DATA
-            // If MeshData is used, NativeArrays are passed to user and
-            // it becomes their responsibility to properly Dispose them.
-            bonesPerVertex.Dispose();
-            boneWeights.Dispose();
-#endif
             boneIndexMap = null;
             boneWeightMap = null;
         }
@@ -753,7 +659,7 @@ namespace Draco {
         // data_type. On input, data must be null. The returned data must be
         // released with ReleaseDracoData.
         [DllImport (DRACODEC_UNITY_LIB)] unsafe static extern bool GetAttributeData(
-            DracoMesh* mesh, DracoAttribute* attr, DracoData**data, bool flip);
+            DracoMesh* mesh, DracoAttribute* attr, DracoData**data, bool flip, int component_stride);
 
         abstract class AttributeMapBase : IComparable<AttributeMapBase> {
             readonly public VertexAttribute attribute;
@@ -794,8 +700,37 @@ namespace Draco {
                 this.convertSpace = convertSpace;
             }
 
-            public override int numComponents => dracoAttribute->numComponents;
-            public override int elementSize => DataTypeSize((DataType)dracoAttribute->dataType) * dracoAttribute->numComponents;
+            /// <summary>
+            /// Unity specifies that attribute data size must be divisible by 4.  
+            /// This value may contain an additional pad to meet this requirement.  
+            /// </summary>
+            public override int numComponents
+            {
+                get
+                {
+                    int dracoElemSize = DataTypeSize((DataType)dracoAttribute->dataType) * dracoAttribute->numComponents;
+
+                    if (dracoElemSize % 4 == 0)
+                    {
+                        return dracoAttribute->numComponents;
+                    }
+                    else
+                    {
+                        // Pad such that element size is divisible by 4.
+                        int padBytes = 4 - dracoElemSize % 4;
+                        int padComponents = padBytes / DataTypeSize((DataType)dracoAttribute->dataType);
+                        return dracoAttribute->numComponents + padComponents;
+                    }
+                }
+            }
+
+            public override int elementSize
+            {
+                get
+                {
+                    return numComponents * DataTypeSize((DataType)dracoAttribute->dataType);
+                }
+            }
 
             public override void Dispose() {
                 var tmp = dracoAttribute;
@@ -804,7 +739,6 @@ namespace Draco {
             }
         }
 
-        
         class CalculatedAttributeMap : AttributeMapBase {
             public int m_numComponents;
             public int m_elementSize;
@@ -817,34 +751,6 @@ namespace Draco {
             public override int numComponents => m_numComponents;
             public override int elementSize => m_elementSize;
         }
-
-#if !DRACO_MESH_DATA
-        abstract class NativeIndexBufferBase : IDisposable {
-            public abstract void* unsafePtr {get;}
-            public abstract int Length {get;}
-            public abstract void SetMeshIndexBufferData(Mesh mesh);
-            public abstract void Dispose();
-        }
-        
-        class NativeIndexBuffer<T> : NativeIndexBufferBase where T : struct {
-            NativeArray<T> indices;
-
-            public NativeIndexBuffer(int length, Allocator allocator) {
-                indices = new NativeArray<T>(length, allocator, NativeArrayOptions.UninitializedMemory);
-            }
-
-            public override void* unsafePtr => indices.GetUnsafePtr();
-            public override int Length => indices.Length;
-            
-            public override void SetMeshIndexBufferData(Mesh mesh) {
-                mesh.SetIndexBufferData( indices, 0, 0, indices.Length);
-            }
-
-            public override void Dispose() {
-                indices.Dispose();
-            }
-        }
-#endif
 
         [BurstCompile]
         struct DecodeJob : IJob {
@@ -906,13 +812,7 @@ namespace Draco {
             public bool flip;
             [ReadOnly]
             public DataType dataType; 
-#if DRACO_MESH_DATA
             public Mesh.MeshData mesh;
-#else
-            [NativeDisableUnsafePtrRestriction]
-            public void* indicesPtr;
-            public int indicesLength;
-#endif
 
             public void Execute() {
                 if (result[0]<0) {
@@ -922,7 +822,6 @@ namespace Draco {
                 if (dracoMesh->isPointCloud) {
                     return;
                 }
-#if DRACO_MESH_DATA
                 void* indicesPtr;
                 int indicesLength;
 
@@ -943,7 +842,6 @@ namespace Draco {
                         result[0] = -1;
                         return;
                 }
-#endif
                 GetMeshIndices(dracoMesh, dataType, indicesPtr, indicesLength, flip);
             }
         }
@@ -963,28 +861,23 @@ namespace Draco {
             [ReadOnly]
             public bool flip;
 
-#if DRACO_MESH_DATA
+            [ReadOnly]
+            public int componentStride;
+
             public Mesh.MeshData mesh;
             [ReadOnly]
             public int streamIndex;
-#else
-            [WriteOnly]
-            [NativeDisableUnsafePtrRestriction]
-            public byte* dstPtr;
-        
-#endif
+
             public void Execute() {
                 if (result[0]<0) {
                     return;
                 }
                 var dracoMesh = (DracoMesh*) dracoTempResources[meshPtrIndex];
                 DracoData* data = null;
-                GetAttributeData(dracoMesh, attribute, &data, flip);
-                var elementSize = DataTypeSize((DataType)data->dataType) * attribute->numComponents;
-#if DRACO_MESH_DATA
+                GetAttributeData(dracoMesh, attribute, &data, flip, componentStride);
+                var elementSize = DataTypeSize((DataType)data->dataType) * componentStride;
                 var dst = mesh.GetVertexData<byte>(streamIndex);
                 var dstPtr = dst.GetUnsafePtr();
-#endif
                 UnsafeUtility.MemCpy(dstPtr, (void*)data->data, elementSize*dracoMesh->numVertices);
                 ReleaseDracoData(&data);
             }
@@ -1008,29 +901,27 @@ namespace Draco {
             [ReadOnly]
             public bool flip;
 
-#if DRACO_MESH_DATA
+            [ReadOnly]
+            public int componentStride;
+
             public Mesh.MeshData mesh;
+            
             [ReadOnly]
             public int streamIndex;
+            
             [ReadOnly]
             public int offset;
-#else
-            [WriteOnly]
-            [NativeDisableUnsafePtrRestriction]
-            public byte* dstPtr;
-#endif
+
             public void Execute() {
                 if (result[0]<0) {
                     return;
                 }
                 var dracoMesh = (DracoMesh*) dracoTempResources[meshPtrIndex];
                 DracoData* data = null;
-                GetAttributeData(dracoMesh, attribute, &data, flip);
-                var elementSize = DataTypeSize((DataType)data->dataType) * attribute->numComponents;
-#if DRACO_MESH_DATA
+                GetAttributeData(dracoMesh, attribute, &data, flip, componentStride);
+                var elementSize = DataTypeSize((DataType)data->dataType) * componentStride;
                 var dst = mesh.GetVertexData<byte>(streamIndex);
                 var dstPtr =  ((byte*)dst.GetUnsafePtr()) + offset;
-#endif
                 for (var v = 0; v < dracoMesh->numVertices; v++) {
                     UnsafeUtility.MemCpy(dstPtr+(stride*v), ((byte*)data->data)+(elementSize*v), elementSize);
                 }
@@ -1071,12 +962,12 @@ namespace Draco {
                 var dracoMesh = (DracoMesh*) dracoTempResources[meshPtrIndex];
 
                 DracoData* indicesData = null;
-                GetAttributeData(dracoMesh, indicesAttribute, &indicesData, false);
+                GetAttributeData(dracoMesh, indicesAttribute, &indicesData, false, indicesAttribute->numComponents);
                 var indicesDataType = (DataType)indicesData->dataType;
                 var indexSize = DataTypeSize((DataType)indicesData->dataType) * indicesAttribute->numComponents;
                 
                 DracoData* weightsData = null;
-                GetAttributeData(dracoMesh, weightsAttribute, &weightsData, false);
+                GetAttributeData(dracoMesh, weightsAttribute, &weightsData, false, weightsAttribute->numComponents);
                 var weightSize = DataTypeSize((DataType)weightsData->dataType) * weightsAttribute->numComponents;
                 
                 for (var v = 0; v < dracoMesh->numVertices; v++) {